issue/596 - 将functional.py中的函数，拆成functional文件夹中的函数
