+++ python/infinicore/nn/functional/__init__.py
@@ -1,0 +1,13 @@
+from .causal_softmax import causal_softmax
+from .random_sample import random_sample
+from .rms_norm import rms_norm
+from .silu import silu
+from .swiglu import swiglu
+
+__all__ = [
+    "causal_softmax",
+    "random_sample",
+    "rms_norm",
+    "silu",
+    "swiglu",
+]
